+++ docs/examples/CommonForms/SignUpSchema.js
@@ -42,7 +42,7 @@ function createValidator(schema) {
     validator(model);
 
     if (validator.errors && validator.errors.length) {
-      throw { details: validator.errors };
+      return { details: validator.errors };
     }
   };
 }