+++ docs/examples/CustomFields/RangeFieldSchema.js
@@ -25,7 +25,7 @@ function createValidator(schema) {
     validator(model);
 
     if (validator.errors && validator.errors.length) {
-      throw { details: validator.errors };
+      return { details: validator.errors };
     }
   };
 }