+++ docs/examples/CustomFields/SwapFieldSchema.js
@@ -20,7 +20,7 @@ function createValidator(schema) {
     validator(model);
 
     if (validator.errors && validator.errors.length) {
-      throw { details: validator.errors };
+      return { details: validator.errors };
     }
   };
 }